update map.js - not working
i keep getting an error that d3.queue is not a function(…)
i have a second file with placeholder data that does run the viz. i am uploading that file into the js file for reference, in order to make that work, you have to "turn off the load data, map, and scatterplot file. not sure what's going on.

still need to figure out how to connect the map to scatterplot